+++ src/cadqueryhelper/shape/__init__.py
@@ -20,3 +20,4 @@ from .chevron import chevron
 from .lightning import lightning
 from .cornerJoin import corner_join
 from .pinwheel import make_circular_points, interweave_lists, pinwheel
+from .trapezoid import trapezoid